rename mock data file for clarity
