+++ frontend/cypress/e2e/subEstablishmentNotDataOwner/EditUser/homePage.spec.cy.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-undef */
 /// <reference types="cypress" />
-import { SubEstablishmentNotDataOwner } from '../../../support/establishmentData';
+import { SubEstablishmentNotDataOwner } from '../../../support/mockEstablishmentData';
 import { onHomePage } from '../../../support/page_objects/onHomePage';
 
 describe('Sub home page as edit user where parent owns the data', () => {
